Add a "get dirty pages list for migration" functionality
A recursive vm_radix tree walk is used to iterate through all of the guest
pages. For each page, the vm_page_test_dirty() function is called to
check whether the page was dirtied and the dirty flag was not yet set.

The uint8_t *page_list must be allocated when calling the ioctl with the
proper value.

Signed-off-by: Elena Mihailescu <elenamihailescu22@gmail.com>